Extract logic from withExternalDataPropertyOptions into selector
